// src/pages/models/dadosUsuario.ts
export interface CadastroDTO{
    login_usuario: string;
    nome: string;
    senha: string;
    confirma_senha: string;
    cpf: string;
    faixa_salarial: string;
    data_nascimento: string;
    rua: string;
    complemento: string;
    bairro: string;
    cidade: string;
    cep: string;
    estado: string;
}